Fix incorrect broadcast address
On at least some linux systems, the broadcast address is unset
(0.0.0.0). Firstly, this is incorrect. Secondly it is not allowed
per RFC1122 3.2.1.3 (a):
https://www.rfc-editor.org/rfc/rfc1122#section-3

The broadcast address is computed from the ip address, and net mask.
broadcast = ip | ~net_mask

This remains the correct vale of the broadcast address regardless of
what SIOCGIFBRDADDR returns. As it will be garbage if no one ever set
it (with SIOCSIFBRDADDR perhaps).

This issue was exposed when attempting to use the loopback network
interface. I've tried a few different distros, and on all of them the
broadcast mask was unset (0.0.0.0). However, on all of them the
following route was set (127.255.255.255 being the broadcast address for
127.0.0.1/8) and resulted in broadcasts being tranmistted and received
properly:

$ ip route get 127.255.255.255
broadcast 127.255.255.255 dev lo table local src 127.0.0.1 uid 1000
    cache <local,brd>