Add option to reduce latency due to TCP delayed ACK
Default in linux is ~40ms. This can and does introduce latency in the
data stream. Protocols sensitive to this type of timing result in
timeouts.

Setting TCP_QUICKACK in sock options disables the delayed ACK. The catch
is that it resets itself after every send/recv - so it needs to be
reenabled upon each call.